[Test] wolfssl key gen.

+++ source/laplace/engine/e_scheduler.cpp
@@ -68,13 +68,15 @@ namespace laplace::engine {
     auto count = thread_count;
 
     if (count < 0) {
-      verb(fmt("Scheduler: Invalid thread count %td (set to 0).", count));
+      verb(fmt("Scheduler: Invalid thread count %d (set to 0).",
+               static_cast<int>(count)));
       count = 0;
     }
 
     if (count > thread_count_limit) {
-      verb(fmt("Scheduler: Invalid thread count %td (max %td).",
-               count, thread_count_limit));
+      verb(fmt("Scheduler: Invalid thread count %d (max %d).",
+               static_cast<int>(count),
+               static_cast<int>(thread_count_limit)));
       count = thread_count_limit;
     }
